requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
-e git+https://github.com/hankadler/python-args2fields@f3cd310ddd7a486d36268e2df982775294452394#egg=args2fields
certifi==2022.6.15; python_version >= '3.6'
charset-normalizer==2.1.1; python_version >= '3.6'
cycler==0.11.0; python_version >= '3.6'
fonttools==4.37.1; python_version >= '3.7'
idna==3.3; python_version >= '3.5'
kiwisolver==1.4.4; python_version >= '3.7'
lxml==4.9.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
matplotlib==3.5.3
multitasking==0.0.11
numpy==1.23.2
packaging==21.3; python_version >= '3.6'
pandas==1.4.4
pillow==9.2.0; python_version >= '3.7'
pyparsing==3.0.9; python_full_version >= '3.6.8'
pyqt5-qt5==5.15.2
pyqt5-sip==12.11.0; python_version >= '3.7'
pyqt5==5.15.7
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2022.2.1
requests==2.28.1; python_version >= '3.7' and python_version < '4'
scipy==1.9.1
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.26.12; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5' and python_version < '4'
yfinance==0.1.74
